+++ app/src/main/java/com/zype/android/ui/player/PlayerViewModel.java
@@ -52,12 +52,15 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
 
     private MutableLiveData<List<PlayerMode>> availablePlayerModes;
     private MutableLiveData<PlayerMode> playerMode;
-    private MutableLiveData<String> playerUrl;
+    private MutableLiveData<String> playerUrl = new MutableLiveData<>();
+    private MutableLiveData<Boolean> isTrailer = new MutableLiveData<>();
     private MutableLiveData<Integer> playbackState = new MutableLiveData<>();
     private MutableLiveData<Error> error = new MutableLiveData<>();
 
     private String videoId;
     private String playlistId;
+    private String trailerVideoId;
+    private String trailerUrl;
 
     private List<AdSchedule> adSchedule;
     private AnalyticBeacon analyticBeacon;
@@ -65,8 +68,8 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
 
     private long playbackPosition = 0;
     private boolean isPlaybackPositionRestored;
-    private boolean isUrlLoaded = false;
     private boolean inBackground = false;
+
     private static final String APP_BUNDLE = "app_bundle";
     private static final String APP_DOMAIN = "app_domain";
     private static final String APP_ID = "app_id";
@@ -77,6 +80,7 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
     private static final String DEVICE_MODEL = "device_model";
     private static final String UUID = "uuid";
     private static final String VPI = "vpi";
+
     public enum PlayerMode {
         AUDIO,
         VIDEO
@@ -109,8 +113,10 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
         oldApi.subscribe(this);
 
         availablePlayerModes = new MutableLiveData<>();
-        availablePlayerModes.setValue(new ArrayList<PlayerMode>());
+        availablePlayerModes.setValue(new ArrayList<>());
         playerMode = new MutableLiveData<>();
+
+        isTrailer.setValue(false);
     }
 
     @Override
@@ -126,6 +132,9 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
         Video video = repo.getVideoSync(videoId);
         playbackPosition = video.playTime;
         isPlaybackPositionRestored = false;
+
+        isTrailer.setValue(false);
+        trailerVideoId = null;
 
         updateAvailablePlayerModes();
         if (mediaType != null || isMediaTypeAvailable(mediaType)) {
@@ -192,6 +201,9 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
     }
 
     public void onPlaybackStarted() {
+        if (isTrailer.getValue()) {
+            return;
+        }
         if(!TextUtils.isEmpty(videoId)) {
             Video video = repo.getVideoSync(videoId);
 
@@ -203,6 +215,9 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
     }
 
     public void onPlaybackFinished() {
+        if (isTrailer.getValue()) {
+            return;
+        }
         if(!TextUtils.isEmpty(videoId)) {
             Video video = repo.getVideoSync(videoId);
 
@@ -244,15 +259,20 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
 
     // Player url
 
-    public MutableLiveData<String> getPlayerUrl() {
+    public LiveData<String> getPlayerUrl() {
         if (playerUrl == null) {
             playerUrl = new MutableLiveData<>();
         }
-        Video video = repo.getVideoSync(videoId);
-        video.playerAudioUrl = null;
-        video.playerVideoUrl = null;
-        updatePlayerUrl(video);
-        loadPlayer();
+        if (isTrailer.getValue()) {
+            playerUrl.setValue(trailerUrl);
+        }
+        else {
+            Video video = repo.getVideoSync(videoId);
+            video.playerAudioUrl = null;
+            video.playerVideoUrl = null;
+            updatePlayerUrl(video);
+            loadPlayer();
+        }
         return playerUrl;
     }
 
@@ -294,7 +314,7 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
 
     // Player mode
 
-    public MutableLiveData<PlayerMode> getPlayerMode() {
+    public LiveData<PlayerMode> getPlayerMode() {
         return playerMode;
     }
 
@@ -309,7 +329,7 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
         }
     }
 
-    public MutableLiveData<List<PlayerMode>> getAvailablePlayerModes() {
+    public LiveData<List<PlayerMode>> getAvailablePlayerModes() {
         return availablePlayerModes;
     }
 
@@ -409,6 +429,26 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
         return error;
     }
 
+    // Trailer
+
+    public LiveData<Boolean> isTrailer() {
+        return isTrailer;
+    }
+
+    public void setTrailerVideoId(String trailerVideoId) {
+        this.trailerVideoId = trailerVideoId;
+        if (TextUtils.isEmpty(trailerVideoId)) {
+            isTrailer.setValue(false);
+            trailerUrl = null;
+            updatePlayerUrl(repo.getVideoSync(videoId));
+        }
+        else {
+            isTrailer.setValue(true);
+            playerUrl.setValue(trailerUrl);
+            loadPlayer();
+        }
+    }
+
     //
 
     public void loadPlayer() {
@@ -438,6 +478,13 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
                 // Take first source in the list
                 String url = files.get(0).url;
                 Logger.d("loadVideoPlayer()::onCompleted(): url=" + url);
+
+                // In play trailer mode just update player url
+                if (isTrailer.getValue()) {
+                    trailerUrl = url;
+                    playerUrl.setValue(trailerUrl);
+                    return;
+                }
 
                 // Ad tags
                 repo.deleteAdSchedule(videoId);
@@ -521,9 +568,16 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
     }
 
     private void loadVideoPlayer(String accessToken, String uuid) {
-        api.getPlayer(videoId, false, accessToken, uuid,
-                getApplication().getString(R.string.app_name) + "/" + BuildConfig.VERSION_NAME,
-                createVideoPlayerListener(accessToken, uuid),getValues());
+        if (isTrailer.getValue()) {
+            api.getPlayer(trailerVideoId, false, accessToken, uuid,
+                    getApplication().getString(R.string.app_name) + "/" + BuildConfig.VERSION_NAME,
+                    createVideoPlayerListener(accessToken, uuid), getValues());
+        }
+        else {
+            api.getPlayer(videoId, false, accessToken, uuid,
+                    getApplication().getString(R.string.app_name) + "/" + BuildConfig.VERSION_NAME,
+                    createVideoPlayerListener(accessToken, uuid), getValues());
+        }
     }
 
     IZypeApiListener createAudioPlayerListener(String accessToken, String uuid) {
@@ -561,7 +615,7 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
     private void loadAudioPlayer(String accessToken, String uuid) {
         api.getPlayer(videoId, true, accessToken, uuid,
                 getApplication().getString(R.string.app_name) + "/" + BuildConfig.VERSION_NAME,
-                createAudioPlayerListener(accessToken, uuid),getValues());
+                createAudioPlayerListener(accessToken, uuid), getValues());
     }
 
     // 'CustomPlayer.InfoListener' implementation
@@ -625,7 +679,7 @@ public class PlayerViewModel extends AndroidViewModel implements CustomPlayer.In
             }
             else {
                 result = new HlsMediaSource.Factory(dataSourceFactory)
-                        .setAllowChunklessPreparation(true)
+//                        .setAllowChunklessPreparation(true)
                         .createMediaSource(Uri.parse(contentUri));
             }
         }
